Member CRUD and API

+++ src/airtext/crud/group_contact.py
@@ -1,4 +1,4 @@
-from airtext.crud.base import DatabaseMixin
+from airtext.crud.base import DatabaseMixin, Response
 from airtext.models.group_contact import GroupContact
 
 
@@ -13,7 +13,11 @@ class GroupContactAPI(DatabaseMixin):
             session.refresh(group_contact)
             session.commit()
 
-        return group_contact
+        return Response(
+            text="Successfully created group contact.",
+            body=group_contact.to_dict(),
+            error=False
+        )
 
     def get_by_group_id(self, group_id: int):
         with self.database() as session:
@@ -29,7 +33,19 @@ class GroupContactAPI(DatabaseMixin):
                 )
                 .first()
             )
+
+            if not group_contact:
+                return Reponse(
+                    text="Group contact not found.",
+                    body={},
+                    error=True,
+                )
+
             session.delete(group_contact)
             session.commit()
 
-        return
+        return Response(
+            text="Successfully removed contact from group.",
+            body={},
+            error=False,
+        )
